+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - ✨(api) update mailboxes #934
+- ✨(api) give update rights to domain viewer on own mailbox #934
 
 ### Changed